test(scanner): move overlay index-masking tests out of scanner.rs
scanner.rs had grown past 14.5k lines. Relocate the self-contained
overlay_index_masking test module (~1000 lines) into the established
dataset/tests/ directory as dataset_overlay_index_masking.rs. The tests
use only crate-public APIs, so no visibility changes were needed.
