+++ rust/lance/src/dataset/tests/mod.rs
@@ -11,6 +11,7 @@ mod dataset_index;
 mod dataset_io;
 mod dataset_merge_update;
 mod dataset_migrations;
+mod dataset_overlay_index_masking;
 mod dataset_scanner;
 mod dataset_schema_evolution;
 mod dataset_transactions;
